# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TAPER",
        "version": "0.1"
    },
    "paths": {
...
        "/api/users/auth/phone": {
            "post": {
                "tags": [
                    "USERS"
                ],
                "summary": "Авторизация по дозвону",
                "operationId": "bbf5ccd59c125881cfe9aa1d96483a63",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "телефон пользователя",
                                        "type": "string",
                                        "example": "89999997777"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/users/auth/check-sms-code": {
            "post": {
                "tags": [
                    "USERS"
                ],
                "summary": "Авторизация по коду из sms или дозвона",
                "operationId": "aea3d014fcb3159d7e87eafb58ba63b1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "телефон пользователя",
                                        "type": "string",
                                        "example": "89999997777"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/users/auth": {
            "post": {
                "tags": [
                    "USERS"
                ],
                "summary": "Авторизация по коду из sms или дозвона",
                "operationId": "c6bb6cc5c5d993d58287f607c4d0398b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Логин пользователя",
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "password": {
                                        "description": "пароль пользователя",
                                        "type": "string",
                                        "example": "qwerty"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/users/auth/check-token": {
            "get": {
                "tags": [
                    "USERS"
                ],
                "summary": "Проверка токена",
                "operationId": "93514d8fd34f20f03cd7f18ab861cb24",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}